+++ hipd/hipd.c
@@ -138,7 +138,7 @@ int hip_use_userspace_ipsec = 0;
 int esp_prot_active;
 int esp_prot_num_transforms = 0;
 uint8_t esp_prot_transforms[MAX_NUM_ESP_PROT_TFMS];
-int esp_prot_num_parallel_hchains = 0;
+long esp_prot_num_parallel_hchains = 0;
 
 int hip_shotgun_status = SO_HIP_SHOTGUN_ON;
